requirements.txt
PyYAML>=6.0
Jinja2>=3.0.3
setuptools>=57.0.0